# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for User Application",
        "title": "OfsetControl",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/control/check-rek": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает вывод команды `ls -lah` для каталога /var/www/reklama.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение списка файлов и прав доступа в каталоге /var/www/reklama",
                "responses": {}
            }
        },
        "/control/check-usb-dir": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет команды `ls -lah /media/user` и `lsusb` для отображения содержимого USB-директории и списка подключённых USB-устройств.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Проверка содержимого /media/user и списка USB-устройств",
                "responses": {}
            }
        },
        "/control/correct-rek": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Устанавливает правильные права доступа и владельца для файлов в каталоге /var/www/reklama.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Исправление прав доступа для каталога /var/www/reklama",
                "responses": {}
            }
        },
        "/control/correct-usb-dir": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Проверяет существование каталога, проверяет его файловую систему (должна быть vfat) и удаляет все файлы в нём.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Очистка каталога /media/user/",
                "responses": {}
            }
        },
        "/control/get-log-bot": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает логи сервиса @bot за последние 2 дня",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение логов бота",
                "responses": {}
            }
        },
        "/control/get-log-prod/{date}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает содержимое лог-файла за выбранную дату",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение логов продакшена",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Дата лога (1 - сегодня, 2 - вчера, 3 - позавчера)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/control/get-log-reboots": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает вывод команды `journalctl` для получения логов, связанных с перезагрузкой системы, начиная с указанной даты.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение логов перезагрузок системы",
                "responses": {}
            }
        },
        "/control/get-lpf-error": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает последние 100 строк из файла лога ошибок LPF (`/opt/offset/log/lpf/error.log`).",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение последних 100 строк ошибок LPF",
                "responses": {}
            }
        },
        "/control/get-lpf-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает последние 100 строк из лог файла LPF и статус очереди печати (используя команду `lpq`).",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Получение информации из лог файла LPF и статуса LPQ",
                "responses": {}
            }
        },
        "/control/order-clean": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Очищает очередь печати, отменяет все задания, включает принтер и перезапускает службу CUPS.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Очистка очереди печати",
                "responses": {}
            }
        },
        "/control/print-test": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Запускает тестовую печать с помощью команды lp, отправляя команду на принтер.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Запуск тестовой печати",
                "responses": {}
            }
        },
        "/control/restart-bot": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Перезапускает службу бота с помощью команды systemctl restart offset@bot.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Перезапуск службы бота",
                "responses": {}
            }
        },
        "/control/run-scanner-test": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Запускает тест сканера с использованием команды `scanimage -T`. Возвращает результат или ошибку в зависимости от выполнения команды.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Выполнение теста сканера",
                "responses": {}
            }
        },
        "/control/run-speedtest": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Запускает тест скорости интернета с использованием команды `speedtest`. Возвращает результат или ошибку в зависимости от выполнения команды.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Выполнение теста скорости интернета",
                "responses": {}
            }
        },
        "/control/run-speedtest-secure": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Запускает защищенную версию теста скорости интернета с использованием команды `speedtest --secure`. Возвращает результат или ошибку в зависимости от выполнения команды.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "control"
                ],
                "summary": "Выполнение теста скорости интернета (защищенная версия)",
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}